spline generation i think?

+++ ftc_app/control/src/main/java/wen/control/Coordinate.java
@@ -2,10 +2,10 @@ package wen.control;
 
 public class Coordinate {
 
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public Coordinate(float x, float y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
